# storage/api-docs/swagger_docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMS Proxy API",
        "description": "API для проксирования запросов к postback-sms.com. По-умолчанию используется токен из ТЗ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://php.mshome.net:8000",
            "description": "Локальный сервер"
        },
        {
            "url": "https://postback-sms.com/",
            "description": "Удалённый сервер"
        }
    ],
    "paths": {
        "/api/get-number": {
            "get": {
                "summary": "Получить номер телефона. По-умолчанию используется токен из ТЗ",
                "operationId": "ba9294b577dd9ad5ac543b455bc34677",
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "service",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rent_time",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "number": {
                                            "type": "string"
                                        },
                                        "activation": {
                                            "type": "string"
                                        },
                                        "cost": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get-sms": {
            "get": {
                "summary": "Получить SMS по activation ID. По-умолчанию используется токен из ТЗ",
                "operationId": "c9d3cd8a5870438547ce881aa9e14ba2",
                "parameters": [
                    {
                        "name": "activation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "sms": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cancel-number": {
            "get": {
                "summary": "Отменить номер. По-умолчанию используется токен из ТЗ",
                "operationId": "ee64678e6cf5d9b98aa80d7eec5017f5",
                "parameters": [
                    {
                        "name": "activation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Номер отменен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "activation": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get-status": {
            "get": {
                "summary": "Получить статус активации. По-умолчанию используется токен из ТЗ",
                "operationId": "546a65109adefbc15f446011c92bbdd7",
                "parameters": [
                    {
                        "name": "activation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Текущий статус активации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "count_sms": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}